MesoHOPS v1.00
This is the initial release of the MesoHOPS code which contains code for both the linear and normalized non-linear Hierarchy of Pure States (HOPS) equations. In addition, it has an adaptive HOPS (adHOPS) algorithm which works with the normalized non-linear HOPS code. Preliminary user guide has been included in the documentation.

NOTE: This version has a known bug in the HopsBasis class that does not correctly treat negative elements in the Hamiltonian.