MesoHops v1.3
This commit is a major update over the previous version of MesoHOPS including:
1. The implementation of an adaptive algorithm that works for arbitrary diagonal L-operators. In practice, this means the adHOPS can now be used for multi-particle dynamics!
2. A substantial speed improvement (3-8x faster for large systems)
3. A modest further improvement in memory when managing a large number of auxiliary wave functions
4. Some behind the scenes simplifications in the code design that will allow enable future generalizations
5. The noise class has been condensed and now has some additional options for how the uncorrelated noise is initially generated

Starting with this commit, we are implementing levels of testing. For most purposes, pytest -level 1 is sufficient to ensure the code is behaving as expected. pytest -level 2 invokes a much slower set of tests that ensures the random number generators are performing within expected tolerances.