MesoHOPS v1.3.1
This update provides a number of quality-of-life improvements (e.g. a larger variety of ways to define the noise trajectory) in the code base and fixes one problem with memory scaling introduced by CPU efficient (but linear scaling memory) algorithm for indexing auxiliary wave functions. The code is also more robust against some errors that were arising in edge-cases.

Behind the scenes, a number of code updates have provided a more consistent set of naming conventions.